local_market/requirements.txt
Babel==2.6.0
Django==1.11.17
django-countries==5.3.2
django-custom-user==0.7
django-model-utils==3.1.2
django-phonenumber-field==2.0.1
phonenumbers==8.10.2
pytz==2018.7
six==1.12.0
swapper==1.1.0
tqdm==4.28.1
